all validated

+++ chapters/5.11.c
@@ -1426,8 +1426,8 @@ void Dynamics(double stopTime, double timeStep)
 int main()
 {
     /* parameters */
-    double randomSeed;
-    randomSeed = 1.0;
+    unsigned int randomSeed;
+    randomSeed = 1;
     srand(randomSeed);
 
     typeMasses[1] = 183.84; // for W
