+++ chapters/DLC/5.12_W.c
@@ -1384,8 +1384,8 @@ void ComputeStress_(double stress[6])
 int main()
 {
     /* parameters */
-    double randomSeed;
-    randomSeed = 1.0;
+    unsigned int randomSeed;
+    randomSeed = 1;
     srand(randomSeed);
 
     typeMasses[1] = 183.84; // for W
