+++ chapters/add/5.13.c
@@ -29,8 +29,8 @@ void Dynamics(double stopTime, double timeStep)
 int main()
 {
     /* parameters */
-    double randomSeed;
-    randomSeed = 1.0;
+    unsigned int randomSeed;
+    randomSeed = 1;
     srand(randomSeed);
 
     typeMasses[1] = 20.1797; // for Ne
